Build fix.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@65566 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Audio/DrvHostALSAAudio.cpp
@@ -1182,7 +1182,7 @@ static DECLCALLBACK(int) drvHostALSAAudioStreamPlay(PPDMIHOSTAUDIO pInterface, P
             break;
         }
 
-        size_t cbToRead = RT_MIN(PDMAUDIOPCMPROPS_S2B(&pStreamALSA->Props, cAvail), cbBuf);
+        size_t cbToRead = RT_MIN((unsigned)PDMAUDIOPCMPROPS_S2B(&pStreamALSA->Props, cAvail), cbBuf);
         if (!cbToRead)
             break;
